+++ web/app/regist/components/ActiveResult.tsx
